+++ src/css/news/newsStand/newsList.css
