+++ src/css/news/newsStand/newsTab.css
