+++ src/index.html
@@ -9,8 +9,8 @@
     <link rel="stylesheet" href="css/layout/header.css">
     <link rel="stylesheet" href="css/news/topNews/rolling.css">
     <link rel="stylesheet" href="css/news/topMenu/menu.css">
-    <link rel="stylesheet" href="css/news/newsstand/newsList.css">
-    <link rel="stylesheet" href="css/news/newsstand/newsTab.css">
+    <link rel="stylesheet" href="css/news/newsStand/newsList.css">
+    <link rel="stylesheet" href="css/news/newsStand/newsTab.css">
 </head>
 <body>
 <div class="container">
